+++ web3lib/src/commonMain/kotlin/com/sonsofcrypto/web3lib/formatters/CurrencyFormatter.kt
@@ -1,8 +1,8 @@
 package com.sonsofcrypto.web3lib.formatters
 
-import com.sonsofcrypto.web3lib.formatters.Formatters.Output
-import com.sonsofcrypto.web3lib.formatters.Formatters.Output.Normal
-import com.sonsofcrypto.web3lib.formatters.Formatters.Style
+import com.sonsofcrypto.web3lib.formatters.Formater.Output
+import com.sonsofcrypto.web3lib.formatters.Formater.Output.Normal
+import com.sonsofcrypto.web3lib.formatters.Formater.Style
 import com.sonsofcrypto.web3lib.types.Currency
 import com.sonsofcrypto.web3lib.utils.BigDec
 import com.sonsofcrypto.web3lib.utils.BigInt