+++ web3lib/src/commonMain/kotlin/com/sonsofcrypto/web3lib/signer/Signer.kt
@@ -134,7 +134,7 @@ abstract class Signer(provider: Provider? = null) {
             chainId = tx.chainId ?: chainId(),
         )
 
-        return tx.copy(gasPrice = unwrappedProvider().estimateGas(tx))
+        return tx.copy(gasLimit = unwrappedProvider().estimateGas(tx))
     }
 
     /** Adds `from` if it does not contain. Throws if `from` != 'address' */